Update throughput with type of total_bytes is __u64

+++ xdp_prog/common_kern_user.h
@@ -46,8 +46,8 @@ typedef struct {
 typedef struct {
     __u64 time_in;
     __u64 proc_time;  /*proc_time += time_out - time_in*/
-    __u32 total_pkts;
-    __u32 total_bytes;
+    __u64 total_pkts;
+    __u64 total_bytes;
 } accounting;
 
 typedef struct svm_weight {
